backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
sqlalchemy
psycopg2-binary
websockets
python-dotenv
asyncpg
greenlet
customtkinter
Pillow
